refactor(cypress): decouple appbuilder UI/API/codemirror commands into commands/appbuilder/ (step 1)
